requirements.txt
contourpy==1.1.1
cycler==0.12.0
dicompyler-core==0.5.6
fonttools==4.43.0
kiwisolver==1.4.5
matplotlib==3.8.0
nibabel==5.1.0
numpy==1.26.0
packaging==23.2
Pillow==10.0.1
pydicom==2.4.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
